# executor: Optimize hash join/agg for `__bin` collator