--- a/pkg/util/collate/gb18030_bin.go
+++ b/pkg/util/collate/gb18030_bin.go
@@ -77,6 +77,11 @@ func (g *gb18030BinCollator) Key(str string) []byte {
 	return g.KeyWithoutTrimRightSpace(truncateTailingSpace(str))
 }
 
+// ImmutableKey implement Collator interface.
+func (g *gb18030BinCollator) ImmutableKey(str string) []byte {
+	return g.Key(str)
+}
+
 // KeyWithoutTrimRightSpace implement Collator interface.
 func (g *gb18030BinCollator) KeyWithoutTrimRightSpace(str string) []byte {
 	buf := make([]byte, 0, len(str))