--- a/pkg/util/collate/gb18030_bin.go
+++ b/pkg/util/collate/gb18030_bin.go
@@ -79,7 +79,7 @@ func (g *gb18030BinCollator) Key(str string) []byte {
 
 // ImmutableKey implement Collator interface.
 func (g *gb18030BinCollator) ImmutableKey(str string) []byte {
-	return g.Key(str)
+	return g.KeyWithoutTrimRightSpace(truncateTailingSpace(str))
 }
 
 // KeyWithoutTrimRightSpace implement Collator interface.