--- a/pkg/util/collate/gb18030_chinese_ci.go
+++ b/pkg/util/collate/gb18030_chinese_ci.go
@@ -73,7 +73,7 @@ func (g *gb18030ChineseCICollator) Key(str string) []byte {
 
 // ImmutableKey implement Collator interface.
 func (g *gb18030ChineseCICollator) ImmutableKey(str string) []byte {
-	return g.Key(str)
+	return g.KeyWithoutTrimRightSpace(truncateTailingSpace(str))
 }
 
 // KeyWithoutTrimRightSpace implement Collator interface.