--- a/pkg/util/collate/general_ci.go
+++ b/pkg/util/collate/general_ci.go
@@ -55,6 +55,11 @@ func (gc *generalCICollator) Key(str string) []byte {
 	return gc.KeyWithoutTrimRightSpace(truncateTailingSpace(str))
 }
 
+// ImmutableKey implement Collator interface.
+func (g *generalCICollator) ImmutableKey(str string) []byte {
+	return g.Key(str)
+}
+
 // KeyWithoutTrimRightSpace implements Collator interface.
 func (*generalCICollator) KeyWithoutTrimRightSpace(str string) []byte {
 	buf := make([]byte, 0, len(str))