--- a/pkg/util/collate/pinyin_tidb_as_cs.go
+++ b/pkg/util/collate/pinyin_tidb_as_cs.go
@@ -28,6 +28,11 @@ func (*zhPinyinTiDBASCSCollator) Key(_ string) []byte {
 	panic("implement me")
 }
 
+// ImmutableKey implement Collator interface.
+func (*zhPinyinTiDBASCSCollator) ImmutableKey(str string) []byte {
+	panic("implement me")
+}
+
 // KeyWithoutTrimRightSpace is not implemented.
 func (*zhPinyinTiDBASCSCollator) KeyWithoutTrimRightSpace(_ string) []byte {
 	panic("implement me")