<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #63507

Problem Summary:

### What changed and how does it work?
1. add `ImmutableKey` to `Collator`, so the caller can use this instead of `Key` if the return value will not be modified.
2. optimize `SerializeKeys` and `HashChunkSelected` for string type by getting collator each chunk instead of each row.
### Check List

Benchmark
```
BenchmarkUtf8mb4Bin_KeyShort-16                   	34731619	        35.26 ns/op	      48 B/op	       1 allocs/op
BenchmarkUtf8mb4Bin_ImmutableKeyShort-16          	476855181	         2.538 ns/op	       0 B/op	       0 allocs/op
BenchmarkUtf8mb4GeneralCI_KeyShort-16             	 4594576	       258.4 ns/op	     128 B/op	       2 allocs/op
BenchmarkUtf8mb4GeneralCI_ImmutableKeyShort-16    	 4685094	       261.7 ns/op	     144 B/op	       2 allocs/op
BenchmarkUtf8mb4UnicodeCI_KeyShort-16             	 5209453	       240.1 ns/op	      96 B/op	       1 allocs/op
BenchmarkUtf8mb4UnicodeCI_ImmutableKeyShort-16    	 4694634	       235.2 ns/op	      96 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_KeyShort-16              	 5230291	       251.4 ns/op	      96 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_ImmutableKeyShort-16     	 5196056	       229.3 ns/op	      80 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_KeyShort-16               	32984313	        34.63 ns/op	      48 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_ImmutableKeyShort-16      	612621256	         2.005 ns/op	       0 B/op	       0 allocs/op
BenchmarkUtf8mb4Bin_KeyMid-16                     	 3097436	       397.5 ns/op	    3072 B/op	       1 allocs/op
BenchmarkUtf8mb4Bin_ImmutableKeyMid-16            	501212491	         2.388 ns/op	       0 B/op	       0 allocs/op
BenchmarkUtf8mb4GeneralCI_KeyMid-16               	   63870	     17622 ns/op	    7168 B/op	       2 allocs/op
BenchmarkUtf8mb4GeneralCI_ImmutableKeyMid-16      	   67890	     18827 ns/op	    7168 B/op	       2 allocs/op
BenchmarkUtf8mb4UnicodeCI_KeyMid-16               	   61974	     19660 ns/op	    6144 B/op	       1 allocs/op
BenchmarkUtf8mb4UnicodeCI_ImmutableKeyMid-16      	   63304	     19058 ns/op	    6144 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_KeyMid-16                	   59450	     18310 ns/op	    6144 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_ImmutableKeyMid-16       	   64184	     18812 ns/op	    6144 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_KeyMid-16                 	 3216708	       392.4 ns/op	    3072 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_ImmutableKeyMid-16        	649398882	         1.863 ns/op	       0 B/op	       0 allocs/op
BenchmarkUtf8mb4Bin_KeyLong-16                    	    4040	    264139 ns/op	 2801671 B/op	       1 allocs/op
BenchmarkUtf8mb4Bin_ImmutableKeyLong-16           	492805200	         2.422 ns/op	       0 B/op	       0 allocs/op
BenchmarkUtf8mb4GeneralCI_KeyLong-16              	      61	  19841427 ns/op	10674194 B/op	       3 allocs/op
BenchmarkUtf8mb4GeneralCI_ImmutableKeyLong-16     	      62	  19808817 ns/op	10674197 B/op	       3 allocs/op
BenchmarkUtf8mb4UnicodeCI_KeyLong-16              	      55	  19803463 ns/op	 5595142 B/op	       1 allocs/op
BenchmarkUtf8mb4UnicodeCI_ImmutableKeyLong-16     	      57	  20020112 ns/op	 5595141 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_KeyLong-16               	      68	  19316250 ns/op	 5595140 B/op	       1 allocs/op
BenchmarkUtf8mb40900AICI_ImmutableKeyLong-16      	      58	  19642420 ns/op	 5595139 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_KeyLong-16                	    4111	    261792 ns/op	 2801672 B/op	       1 allocs/op
BenchmarkUtf8mb40900Bin_ImmutableKeyLong-16       	650685670	         1.862 ns/op	       0 B/op	       0 allocs/op
```

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
Optimize the performance of hash agg/join when string with `xx_bin` collator as the agg/join key
```